add di, add items to db, fix dependencies,

+++ app/src/main/java/com/egorovoy/testtask/data/local/entity/ItemEntity.kt
@@ -12,7 +12,7 @@ data class ItemEntity(
     @ColumnInfo(name = "name")
     val name: String,
     @ColumnInfo(name = "time")
-    val time: Int,
+    val time: Long,
     @ColumnInfo(name = "tags")
     val tags: String,
     @ColumnInfo(name = "amount")
